refactor: split constraint tests in custom folder
